fix: macOS cross-compilation
Removes -march=native and -mtune=native from the podspec.
As these break cross-compilation of the library for x86_64, while on an arm64 mac.
Which is part of the flutter's compilation for macOS.

The Clang error that was thrown was: "error: unknown target CPU 'apple-m1'"
and it only happened on release builds.

Signed-off-by: Dan Printzell <dan@dagg.ai>